syntax cleaning

+++ UI_MatrixCalculator.EXMPL/GUInterface/ReturnEquation.cs
@@ -4,13 +4,13 @@ using System.Windows.Controls;
 
 namespace UI_MatrixCalculator.EXMPL.GUInterface {
     public static class ReturnEquation {
-        public static List<int> NumsFromGrid(Grid grid) {
-            var lstNums = new List<int>();
+        public static List<double> NumsFromGrid(Grid grid) {
+            var lstNums = new List<double>();
             foreach (var element in grid.Children) {
                 if (element.GetType() != typeof(TextBox)) continue;
                 var temp = element as TextBox;
                 if (temp!.Visibility != Visibility.Visible) continue;
-                if (int.TryParse(temp.Text,out var number)) lstNums.Add(number);
+                if (double.TryParse(temp.Text,out var number)) lstNums.Add(number);
             }
             
             return lstNums;
